* main: (48 commits)
  Update release notes.
  Refactor macro _FillValue to NC_FillValue in support of https://github.com/Unidata/netcdf-c/issues/2858
  Remove appveyor config file.
  Add CI for a Windows Runner on Github Actions.
  Add cmake prefix path to appveyor config.
  Attempt to fix zlib-related error in appveyor.
  Correct typo.
  Remove check for libcurl unless it is necessary for required functionality.
  Rename the vendored strlcat symbol
  Fix conversion warnings in libdispatch
  Set flags to avoid warning messages if curl isn't found.
  Use modern cmake nomenclature for curl.
  Fix truncated-format warning in ncgen
  Fix some conversion warnings in ncgen3
  Fix some conversion warnings in ncgen3 generated files
  Regenerate ncgen3
  CMake: Add option to automatically regenerate ncgen3/ncgen
  Skip checking for duplicates if only one element in list
  Change format of backwards-loops
  Silence some conversion warnings in ncgen generated files
  ...

+++ libdispatch/ncbytes.c
@@ -203,7 +203,7 @@ ncbytesremove(NCbytes* bb, unsigned long pos)
     if(bb == NULL) return ncbytesfail();
     if(bb->length <= pos) return ncbytesfail();
     if(pos < (bb->length - 1)) {
-	int copylen = (bb->length - pos) - 1;
+	size_t copylen = (bb->length - pos) - 1;
         memmove(bb->content+pos,bb->content+pos+1,copylen);
     }
     bb->length--;
